sniffer wit new message classes system is stable

+++ com/ankamagames/dofus/network/messages/common/basic/BasicStatMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class BasicStatMessage(NetworkMessage):
     protocolId = 514
-    timeSpent:float
+    timeSpent:int
     statId:int
