+++ com/ankamagames/dofus/network/messages/game/PaginationAnswerAbstractMessage.py
@@ -3,7 +3,7 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class PaginationAnswerAbstractMessage(NetworkMessage):
     protocolId = 2864
-    offset:float
+    offset:int
     count:int
     total:int
