+++ com/ankamagames/dofus/network/messages/game/PaginationRequestAbstractMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class PaginationRequestAbstractMessage(NetworkMessage):
     protocolId = 789
-    offset:float
+    offset:int
     count:int
