+++ com/ankamagames/dofus/network/messages/game/actions/AbstractGameActionMessage.py
@@ -4,5 +4,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class AbstractGameActionMessage(NetworkMessage):
     protocolId = 5037
     actionId:int
-    sourceId:float
+    sourceId:int
