+++ com/ankamagames/dofus/network/messages/game/actions/fight/GameActionFightDropCharacterMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.game.actions.AbstractGameActionMessa
 
 class GameActionFightDropCharacterMessage(AbstractGameActionMessage):
     protocolId = 2160
-    targetId:float
+    targetId:int
     cellId:int
