+++ com/ankamagames/dofus/network/messages/game/actions/fight/GameActionFightLifePointsGainMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.game.actions.AbstractGameActionMessa
 
 class GameActionFightLifePointsGainMessage(AbstractGameActionMessage):
     protocolId = 6777
-    targetId:float
+    targetId:int
     delta:int
