+++ com/ankamagames/dofus/network/messages/game/actions/sequence/SequenceEndMessage.py
@@ -4,6 +4,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class SequenceEndMessage(NetworkMessage):
     protocolId = 5912
     actionId:int
-    authorId:float
+    authorId:int
     sequenceType:int
