+++ com/ankamagames/dofus/network/messages/game/actions/sequence/SequenceStartMessage.py
@@ -4,5 +4,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class SequenceStartMessage(NetworkMessage):
     protocolId = 8598
     sequenceType:int
-    authorId:float
+    authorId:int
