+++ com/ankamagames/dofus/network/messages/game/atlas/compass/CompassUpdatePvpSeekMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.game.atlas.compass.CompassUpdateMess
 
 class CompassUpdatePvpSeekMessage(CompassUpdateMessage):
     protocolId = 5714
-    memberId:float
+    memberId:int
     memberName:str
