+++ com/ankamagames/dofus/network/messages/game/character/alignment/war/effort/AlignmentWarEffortDonatePreviewMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class AlignmentWarEffortDonatePreviewMessage(NetworkMessage):
     protocolId = 2736
-    xp:float
+    xp:int
