+++ com/ankamagames/dofus/network/messages/game/character/alignment/war/effort/AlignmentWarEffortDonateRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class AlignmentWarEffortDonateRequestMessage(NetworkMessage):
     protocolId = 5249
-    donation:float
+    donation:int
