+++ com/ankamagames/dofus/network/messages/game/character/alignment/war/effort/CharacterAlignmentWarEffortProgressionMessage.py
@@ -3,7 +3,7 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class CharacterAlignmentWarEffortProgressionMessage(NetworkMessage):
     protocolId = 7324
-    alignmentWarEffortDailyLimit:float
-    alignmentWarEffortDailyDonation:float
-    alignmentWarEffortPersonalDonation:float
+    alignmentWarEffortDailyLimit:int
+    alignmentWarEffortDailyDonation:int
+    alignmentWarEffortPersonalDonation:int
