+++ com/ankamagames/dofus/network/messages/game/character/deletion/CharacterDeletionRequestMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class CharacterDeletionRequestMessage(NetworkMessage):
     protocolId = 8394
-    characterId:float
+    characterId:int
     secretAnswerHash:str
