+++ com/ankamagames/dofus/network/messages/game/character/replay/CharacterReplayRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class CharacterReplayRequestMessage(NetworkMessage):
     protocolId = 9614
-    characterId:float
+    characterId:int
