+++ com/ankamagames/dofus/network/messages/game/character/stats/CharacterExperienceGainMessage.py
@@ -3,8 +3,8 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class CharacterExperienceGainMessage(NetworkMessage):
     protocolId = 4524
-    experienceCharacter:float
-    experienceMount:float
-    experienceGuild:float
-    experienceIncarnation:float
+    experienceCharacter:int
+    experienceMount:int
+    experienceGuild:int
+    experienceIncarnation:int
