+++ com/ankamagames/dofus/network/messages/game/chat/ChatServerCopyMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.game.chat.ChatAbstractServerMessage 
 
 class ChatServerCopyMessage(ChatAbstractServerMessage):
     protocolId = 5344
-    receiverId:float
+    receiverId:int
     receiverName:str
