+++ com/ankamagames/dofus/network/messages/game/chat/smiley/ChatSmileyMessage.py
@@ -3,7 +3,7 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class ChatSmileyMessage(NetworkMessage):
     protocolId = 5518
-    entityId:float
+    entityId:int
     smileyId:int
     accountId:int
