+++ com/ankamagames/dofus/network/messages/game/context/GameContextReadyMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameContextReadyMessage(NetworkMessage):
     protocolId = 912
-    mapId:float
+    mapId:int
