+++ com/ankamagames/dofus/network/messages/game/context/ShowCellMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class ShowCellMessage(NetworkMessage):
     protocolId = 2286
-    sourceId:float
+    sourceId:int
     cellId:int
