+++ com/ankamagames/dofus/network/messages/game/context/fight/GameFightHumanReadyStateMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameFightHumanReadyStateMessage(NetworkMessage):
     protocolId = 4318
-    characterId:float
+    characterId:int
     isReady:bool
