+++ com/ankamagames/dofus/network/messages/game/context/fight/GameFightLeaveMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameFightLeaveMessage(NetworkMessage):
     protocolId = 4663
-    charId:float
+    charId:int
