+++ com/ankamagames/dofus/network/messages/game/context/fight/GameFightPlacementSwapPositionsCancelledMessage.py
@@ -4,5 +4,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class GameFightPlacementSwapPositionsCancelledMessage(NetworkMessage):
     protocolId = 998
     requestId:int
-    cancellerId:float
+    cancellerId:int
