+++ com/ankamagames/dofus/network/messages/game/context/fight/GameFightPlacementSwapPositionsOfferMessage.py
@@ -4,8 +4,8 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class GameFightPlacementSwapPositionsOfferMessage(NetworkMessage):
     protocolId = 519
     requestId:int
-    requesterId:float
+    requesterId:int
     requesterCellId:int
-    requestedId:float
+    requestedId:int
     requestedCellId:int
