+++ com/ankamagames/dofus/network/messages/game/context/fight/GameFightPlacementSwapPositionsRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.game.context.fight.GameFightPlacemen
 
 class GameFightPlacementSwapPositionsRequestMessage(GameFightPlacementPositionRequestMessage):
     protocolId = 7699
-    requestedId:float
+    requestedId:int
