+++ com/ankamagames/dofus/network/messages/game/context/fight/GameFightRemoveTeamMemberMessage.py
@@ -5,5 +5,5 @@ class GameFightRemoveTeamMemberMessage(NetworkMessage):
     protocolId = 6697
     fightId:int
     teamId:int
-    charId:float
+    charId:int
