+++ com/ankamagames/dofus/network/messages/game/context/fight/GameFightTurnEndMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameFightTurnEndMessage(NetworkMessage):
     protocolId = 4443
-    id:float
+    id:int
