+++ com/ankamagames/dofus/network/messages/game/context/fight/SlaveNoLongerControledMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class SlaveNoLongerControledMessage(NetworkMessage):
     protocolId = 4540
-    masterId:float
-    slaveId:float
+    masterId:int
+    slaveId:int
