+++ com/ankamagames/dofus/network/messages/game/context/mount/PaddockBuyRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class PaddockBuyRequestMessage(NetworkMessage):
     protocolId = 8638
-    proposedPrice:float
+    proposedPrice:int
