+++ com/ankamagames/dofus/network/messages/game/context/roleplay/TeleportOnSameMapMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class TeleportOnSameMapMessage(NetworkMessage):
     protocolId = 9521
-    targetId:float
+    targetId:int
     cellId:int
