+++ com/ankamagames/dofus/network/messages/game/context/roleplay/anomaly/MapComplementaryInformationsAnomalyMessage.py
@@ -4,5 +4,5 @@ from com.ankamagames.dofus.network.messages.game.context.roleplay.MapComplementa
 class MapComplementaryInformationsAnomalyMessage(MapComplementaryInformationsDataMessage):
     protocolId = 6414
     level:int
-    closingTime:float
+    closingTime:int
