+++ com/ankamagames/dofus/network/messages/game/context/roleplay/breach/BreachEnterMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class BreachEnterMessage(NetworkMessage):
     protocolId = 6485
-    owner:float
+    owner:int
