+++ com/ankamagames/dofus/network/messages/game/context/roleplay/breach/meeting/BreachKickRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class BreachKickRequestMessage(NetworkMessage):
     protocolId = 2909
-    target:float
+    target:int
