+++ com/ankamagames/dofus/network/messages/game/context/roleplay/death/GameRolePlayPlayerLifeStatusMessage.py
@@ -4,5 +4,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class GameRolePlayPlayerLifeStatusMessage(NetworkMessage):
     protocolId = 7689
     state:int
-    phenixMapId:float
+    phenixMapId:int
