+++ com/ankamagames/dofus/network/messages/game/context/roleplay/delay/GameRolePlayDelayedActionFinishedMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameRolePlayDelayedActionFinishedMessage(NetworkMessage):
     protocolId = 6062
-    delayedCharacterId:float
+    delayedCharacterId:int
     delayTypeId:int
