+++ com/ankamagames/dofus/network/messages/game/context/roleplay/delay/GameRolePlayDelayedActionMessage.py
@@ -3,7 +3,7 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameRolePlayDelayedActionMessage(NetworkMessage):
     protocolId = 1161
-    delayedCharacterId:float
+    delayedCharacterId:int
     delayTypeId:int
-    delayEndTime:float
+    delayEndTime:int
