+++ com/ankamagames/dofus/network/messages/game/context/roleplay/emote/EmotePlayAbstractMessage.py
@@ -4,5 +4,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class EmotePlayAbstractMessage(NetworkMessage):
     protocolId = 4497
     emoteId:int
-    emoteStartTime:float
+    emoteStartTime:int
