+++ com/ankamagames/dofus/network/messages/game/context/roleplay/emote/EmotePlayMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.game.context.roleplay.emote.EmotePla
 
 class EmotePlayMessage(EmotePlayAbstractMessage):
     protocolId = 9462
-    actorId:float
+    actorId:int
     accountId:int
