+++ com/ankamagames/dofus/network/messages/game/context/roleplay/fight/GameRolePlayAggressionMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameRolePlayAggressionMessage(NetworkMessage):
     protocolId = 2660
-    attackerId:float
-    defenderId:float
+    attackerId:int
+    defenderId:int
