+++ com/ankamagames/dofus/network/messages/game/context/roleplay/fight/GameRolePlayAttackMonsterRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameRolePlayAttackMonsterRequestMessage(NetworkMessage):
     protocolId = 3767
-    monsterGroupId:float
+    monsterGroupId:int
