+++ com/ankamagames/dofus/network/messages/game/context/roleplay/fight/GameRolePlayFightRequestCanceledMessage.py
@@ -4,6 +4,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class GameRolePlayFightRequestCanceledMessage(NetworkMessage):
     protocolId = 4478
     fightId:int
-    sourceId:float
-    targetId:float
+    sourceId:int
+    targetId:int
