+++ com/ankamagames/dofus/network/messages/game/context/roleplay/fight/GameRolePlayMonsterAngryAtPlayerMessage.py
@@ -3,8 +3,8 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameRolePlayMonsterAngryAtPlayerMessage(NetworkMessage):
     protocolId = 465
-    playerId:float
-    monsterGroupId:float
-    angryStartTime:float
-    attackTime:float
+    playerId:int
+    monsterGroupId:int
+    angryStartTime:int
+    attackTime:int
