+++ com/ankamagames/dofus/network/messages/game/context/roleplay/fight/GameRolePlayMonsterNotAngryAtPlayerMessage.py
@@ -3,6 +3,6 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GameRolePlayMonsterNotAngryAtPlayerMessage(NetworkMessage):
     protocolId = 7726
-    playerId:float
-    monsterGroupId:float
+    playerId:int
+    monsterGroupId:int
