+++ com/ankamagames/dofus/network/messages/game/context/roleplay/fight/GameRolePlayPlayerFightFriendlyAnsweredMessage.py
@@ -4,7 +4,7 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 class GameRolePlayPlayerFightFriendlyAnsweredMessage(NetworkMessage):
     protocolId = 5417
     fightId:int
-    sourceId:float
-    targetId:float
+    sourceId:int
+    targetId:int
     accept:bool
