+++ com/ankamagames/dofus/network/messages/game/context/roleplay/havenbag/EnterHavenBagRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class EnterHavenBagRequestMessage(NetworkMessage):
     protocolId = 8214
-    havenBagOwner:float
+    havenBagOwner:int
