+++ com/ankamagames/dofus/network/messages/game/context/roleplay/havenbag/KickHavenBagRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class KickHavenBagRequestMessage(NetworkMessage):
     protocolId = 188
-    guestId:float
+    guestId:int
