+++ com/ankamagames/dofus/network/messages/game/context/roleplay/havenbag/meeting/TeleportHavenBagRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class TeleportHavenBagRequestMessage(NetworkMessage):
     protocolId = 5838
-    guestId:float
+    guestId:int
