+++ com/ankamagames/dofus/network/messages/game/context/roleplay/houses/HouseBuyRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class HouseBuyRequestMessage(NetworkMessage):
     protocolId = 4286
-    proposedPrice:float
+    proposedPrice:int
