+++ com/ankamagames/dofus/network/messages/game/guild/GuildKickRequestMessage.py
@@ -3,5 +3,5 @@ from com.ankamagames.dofus.network.messages.NetworkMessage import NetworkMessage
 
 class GuildKickRequestMessage(NetworkMessage):
     protocolId = 3965
-    kickedId:float
+    kickedId:int
